[Feat] #278 - 알림 권한 확인 로직을 ViewModel로 변경
